+++ forest-client/Forest/obj/x64/Debug/Forest.AssemblyInfo.cs
@@ -14,7 +14,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyDescriptionAttribute("A sample plugin.")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("0.0.0.1")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("0.0.0.1+d7149ce1ea18ac48fca548102632cf50165f8091")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("0.0.0.1+4ae7ce152fcd46360617a06ad8e4daf2bd632d48")]
 [assembly: System.Reflection.AssemblyProductAttribute("Forest")]
 [assembly: System.Reflection.AssemblyTitleAttribute("Forest")]
 [assembly: System.Reflection.AssemblyVersionAttribute("0.0.0.1")]